Extract most of the l10n code to agama-l10n
