+++ rust/agama-l10n/src/error.rs
