+++ rust/agama-l10n/src/helpers.rs
