+++ rust/agama-l10n/src/info.rs
