+++ rust/agama-l10n/src/lib.rs
@@ -18,3 +18,13 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
+mod error;
+pub mod helpers;
+pub mod info;
+mod model;
+mod proposal;
+
+pub use error::LocaleError;
+pub use info::LocaleInfo;
+pub use model::{Keymap, L10n, LocaleEntry, TimezoneEntry};
+pub use proposal::LocalizationProposal;
