+++ rust/agama-l10n/src/model/keyboard.rs
