+++ rust/agama-lib/src/config.rs
@@ -1,4 +1,4 @@
-// Copyright (c) [2024] SUSE LLC
+// Copyright (c) [2025] SUSE LLC
 //
 // All Rights Reserved.
 //
@@ -18,8 +18,5 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-//! Implements support for handling the localization settings
-
-mod settings;
-
-pub use settings::LocalizationSettings;
+mod l10n;
+pub use l10n::LocalizationConfig;
