+++ rust/agama-server/Cargo.toml
@@ -11,6 +11,7 @@ anyhow = "1.0"
 agama-locale-data = { path = "../agama-locale-data" }
 agama-lib = { path = "../agama-lib" }
 agama-utils = { path = "../agama-utils" }
+agama-l10n = { path = "../agama-l10n" }
 zbus = { version = "5", default-features = false, features = ["tokio"] }
 uuid = { version = "1.10.0", features = ["v4"] }
 thiserror = "2.0.12"
