+++ rust/agama-server/src/agama-dbus-server.rs
@@ -18,11 +18,8 @@
 // To contact SUSE LLC about this file by physical or electronic mail, you may
 // find current contact information at www.suse.com.
 
-use agama_server::{
-    l10n::{self, helpers},
-    logs::init_logging,
-    questions,
-};
+use agama_l10n::helpers as l10n_helpers;
+use agama_server::{logs::init_logging, questions};
 
 use agama_lib::connection_to;
 use anyhow::Context;
@@ -33,7 +30,7 @@ const SERVICE_NAME: &str = "org.opensuse.Agama1";
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let locale = helpers::init_locale()?;
+    let locale = l10n_helpers::init_locale()?;
     init_logging().context("Could not initialize the logger")?;
 
     let connection = connection_to(ADDRESS)
