+++ rust/agama-server/src/error.rs
@@ -27,7 +27,6 @@ use axum::{
 use serde_json::json;
 
 use crate::{
-    l10n::LocaleError,
     users::password::PasswordCheckerError,
     web::common::{IssuesServiceError, ProgressServiceError},
 };
@@ -42,8 +41,6 @@ pub enum Error {
     Service(#[from] ServiceError),
     #[error("Questions service error: {0}")]
     Questions(QuestionsError),
-    #[error("Software service error: {0}")]
-    Locale(#[from] LocaleError),
     #[error("Issues service error: {0}")]
     Issues(#[from] IssuesServiceError),
     #[error("Progress service error: {0}")]
