+++ rust/agama-server/src/l10n.rs
@@ -1,4 +1,4 @@
-// Copyright (c) [2024] SUSE LLC
+// Copyright (c) [2025] SUSE LLC
 //
 // All Rights Reserved.
 //
@@ -19,13 +19,4 @@
 // find current contact information at www.suse.com.
 
 mod agent;
-pub mod error;
-pub mod helpers;
-pub mod info;
-mod model;
-
-pub use agama_lib::localization::model::LocaleConfig;
 pub use agent::L10nAgent;
-pub use error::LocaleError;
-pub use info::LocaleInfo;
-pub use model::{Keymap, L10n, LocaleEntry, TimezoneEntry};
